requirements.txt
tensorflow
scikit-learn
matplotlib
seaborn
numpy
Pillow
streamlit
streamlit-webrtc
opencv-python-headless
google-generativeai
